fix(kernel): close occt-wasm in-repo parity gaps (hull, B-spline, divergences)
Three classes of occt-wasm parity failures from the #1136 audit that need
no WASM rebuild:

- hull: the adapter declined hull(shapes) and its gift-wrapping
  computeConvexHullFaces produced wrong hulls (a 10x10x10 box hulled to
  ~333 instead of 1000). Extract the proven QuickHull into a shared
  src/kernel/hullGeometry.ts used by both the occt and occt-wasm adapters;
  occt-wasm now samples surface vertices via tessellate and reconstructs
  with buildSolidFromFaces. Fixes hull and minkowski (which composes hull).
- BSplineToBezier: makeBSpline2d returned a Bezier (so geomType was
  BEZIER_CURVE, not BSPLINE_CURVE) and decomposeBSpline2dToBeziers emitted
  trimmed B-splines. makeBSpline2d now builds a real clamped B-spline;
  decompose implements NURBS Book A5.6 knot-insertion to emit true Beziers.
- Register occt-wasm divergences for raw-`oc` test artifacts (findCurveType
  feeds oc.GeomAbs_* enums; guided/multi-section sweep build spines via raw oc).

occt-wasm parity suite: 27 -> 9 failures. The remaining 9 are native gaps
pending an occt-wasm WASM rebuild (helix occt-wasm#125; uvMesh + compsolid
occt-wasm#126).
